Refactor integrations page to use dedicated components and simplify form handling
